basic functionalities of the app initializes
