+++ lib/l10n/intl_si.arb
@@ -1,0 +1,9 @@
+{
+  "appTitle": "සර්ප ගේම්",
+  "startGame": "ක‍්රීඩාව ආරම්භ කරන්න",
+  "instructions": "ඉගැන්වීම්",
+  "instructionText": "සර්පයා පාලනය සඳහා ස්වයිප් කරන්න. ලකුණු ලබා ගැනීමට ආහාර කන්න. තනිව හෝ බිත්තිවලට වැදීමෙන් වළකින්න!",
+  "playAgain": "නැවත ක්‍රීඩා කරන්න",
+  "scoreText": "ලකුණු: {score}",
+  "highScoreText": "උපරිම ලකුණු: {highScore}"
+}
