+++ lib/l10n/intl_ta.arb
@@ -1,0 +1,9 @@
+{
+  "appTitle": "பாம்பு விளையாட்டு",
+  "startGame": "விளையாட்டை தொடங்கு",
+  "instructions": "வழிமுறைகள்",
+  "instructionText": "பாம்பை கட்டுப்படுத்த ஸ்வைப் செய்யவும். மதிப்பெண்களைப் பெற உணவைச் சாப்பிடவும். உங்களைத் தட்டாதீர்கள் அல்லது சுவர்களுடன் மோதாதீர்கள்!",
+  "playAgain": "மீண்டும் விளையாடு",
+  "scoreText": "மதிப்பெண்கள்: {score}",
+  "highScoreText": "உயர் மதிப்பெண்கள்: {highScore}"
+}
